divide main module into files
