+++ pylev/Levenshtein.py
@@ -1,26 +1,4 @@
-"""
-    pylev (fork by hell03end)
-    =====
-
-    A Python Levenshtein OOP implementation that's not freaking GPL'd.
-
-    Based off the Wikipedia code samples at
-    http://en.wikipedia.org/wiki/Levenshtein_distance.
-    Original version: https://github.com/toastdriven/pylev.
-
-    Usage
-    -----
-
-    Usage is fairly straightforward:
-
-        from pylev import Levenshtein
-        >>> Levenshtein().wf('kitten', 'sitting')
-        3
-"""
-
-__author__ = ('Daniel Lindsley', 'hell03end')
-__version__ = (1, 3, 1)
-__license__ = 'New BSD'
+''' Main Levenshtein class (contain all functions as methods) '''
 
 
 class Levenshtein(object):
@@ -301,18 +279,6 @@ class Levenshtein(object):
                 3
         """
         return self.wfi(string_1, string_2)
-
-
-# for backward-compatibilty with original pylev
-classic_levenshtein = Levenshtein().classic
-recursive_levenshtein = Levenshtein().recursive
-wf_levenshtein = Levenshtein().wf
-wfi_levenshtein = Levenshtein().wfi
-damerau_levenshtein = Levenshtein().damerau
-
-
-# import only Levenshtein class
-__all__ = ("Levenshtein")
 
 
 if __name__ == "__main__":
